Tests/LibWeb/Text/expected/WebSocket/send-out-of-bounds-buffer-source.txt

open
typed array send threw TypeError
data view send threw TypeError
close
